+++ pkg/gardenlet/controller/seed/seed/nginxingress.go
@@ -25,6 +25,7 @@ import (
 	"github.com/gardener/gardener/pkg/operation/botanist/component"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/extensions/dnsrecord"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/nginxingress"
+	seedpkg "github.com/gardener/gardener/pkg/operation/seed"
 	"github.com/gardener/gardener/pkg/utils"
 	"github.com/gardener/gardener/pkg/utils/images"
 	"github.com/gardener/gardener/pkg/utils/imagevector"
@@ -45,7 +46,7 @@ import (
 
 const annotationSeedIngressClass = "seed.gardener.cloud/ingress-class"
 
-func managedIngress(seed *Seed) bool {
+func managedIngress(seed *seedpkg.Seed) bool {
 	return seed.GetInfo().Spec.DNS.Provider != nil && seed.GetInfo().Spec.Ingress != nil && seed.GetInfo().Spec.Ingress.Controller.Kind == v1beta1constants.IngressKindNginx
 }
 
@@ -120,7 +121,7 @@ func getManagedIngressDNSRecord(
 	)
 }
 
-func migrateIngressClassForShootIngresses(ctx context.Context, gardenClient, seedClient client.Client, seed *Seed, newClass string, kubernetesVersion *semver.Version) error {
+func migrateIngressClassForShootIngresses(ctx context.Context, gardenClient, seedClient client.Client, seed *seedpkg.Seed, newClass string, kubernetesVersion *semver.Version) error {
 	if oldClass, ok := seed.GetInfo().Annotations[annotationSeedIngressClass]; ok && oldClass == newClass {
 		return nil
 	}
@@ -201,7 +202,7 @@ func switchIngressClass(ctx context.Context, seedClient client.Client, ingressKe
 	return seedClient.Update(ctx, ingress)
 }
 
-func getConfig(seed *Seed) (map[string]string, error) {
+func getConfig(seed *seedpkg.Seed) (map[string]string, error) {
 	var (
 		defaultConfig = map[string]interface{}{
 			"server-name-hash-bucket-size": "256",
@@ -220,7 +221,7 @@ func getConfig(seed *Seed) (map[string]string, error) {
 }
 
 // ComputeNginxIngressClass returns the IngressClass for the Nginx Ingress controller
-func ComputeNginxIngressClass(seed *Seed, kubernetesVersion *string) (string, error) {
+func ComputeNginxIngressClass(seed *seedpkg.Seed, kubernetesVersion *string) (string, error) {
 	managed := managedIngress(seed)
 
 	if kubernetesVersion == nil {
